api/requirements.txt
Flask==1.1.2
PyYAML==5.4.1
requests==2.25.1
SQLAlchemy==1.3.23
gevent==21.1.2
mysqlclient==2.0.3